rename id80 v2 filenames
