+++ keyboards/idobao/id80/v2/rules.mk
@@ -1,5 +1,5 @@
 # Defalt to the ansi version
-DEFAULT_FOLDER = idobao/id80/v1/ansi
+DEFAULT_FOLDER = idobao/id80/v2/ansi
 
 # MCU name
 MCU = atmega32u4
@@ -10,15 +10,12 @@ BOOTLOADER = atmel-dfu
 # Build Options
 #   change yes to no to disable
 #
-BOOTMAGIC_ENABLE = yes      # Enable Bootmagic Lite
+BOOTMAGIC_ENABLE = yes       # Enable Bootmagic Lite
 MOUSEKEY_ENABLE = yes        # Mouse keys
 EXTRAKEY_ENABLE = yes        # Audio control and System control
 CONSOLE_ENABLE = no          # Console for debug
 COMMAND_ENABLE = no          # Commands for debug and configuration
-NKRO_ENABLE = yes           # Enable N-Key Rollover
+NKRO_ENABLE = yes            # Enable N-Key Rollover
 BACKLIGHT_ENABLE = yes       # Enable keyboard backlight functionality
 RGBLIGHT_ENABLE = yes        # Enable keyboard RGB underglow
 AUDIO_ENABLE = no            # Audio output
-
-# partially generated by KBFirmware JSON to QMK Parser
-# https://noroadsleft.github.io/kbf_qmk_converter/
